requirements.txt
flask
flask_session
flask-socketio
eventlet
requests
websocket-client
python-dotenv
# Usando a biblioteca recomendada para OBS v5+
obsws-python
pygame
pynput
cryptography
# Adicionado para validação de imagem (opcional, mas recomendado, ou use imghdr nativo)
Pillow